+++ src/auth/oauth_device_flow.cpp
@@ -440,7 +440,7 @@ std::string OAuthDeviceFlow::urlEncode(const std::string &value) {
         return value;
     }
 
-    char *encoded = curl_easy_escape(curl, value.c_str(), value.size());
+    char *encoded = curl_easy_escape(curl, value.c_str(), static_cast<int>(value.size()));
     std::string result = {};
     if (encoded) {
         result = encoded;
