+++ src/auth/oauth_pkce_flow.cpp
@@ -430,7 +430,7 @@ std::string OAuthPKCEFlow::urlEncode(const std::string &value) {
         return value;
     }
 
-    char *encoded = curl_easy_escape(curl, value.c_str(), value.size());
+    char *encoded = curl_easy_escape(curl, value.c_str(), static_cast<int>(value.size()));
     std::string result = {};
     if (encoded) {
         result = encoded;
