+++ src/security/usb_admin_authenticator.cpp
@@ -478,7 +478,7 @@ std::optional<USBAdminLicense> USBAdminAuthenticator::loadLicenseFromUSB() const
 // Helper: Base64 decode
 static std::vector<uint8_t> base64Decode(const std::string& encoded) {
     USBAdmin_BIO_ptr b64(BIO_new(BIO_f_base64()));
-    USBAdmin_BIO_ptr bmem(BIO_new_mem_buf(encoded.data(), encoded.size()));
+    USBAdmin_BIO_ptr bmem(BIO_new_mem_buf(encoded.data(), static_cast<int>(encoded.size())));
     BIO* result = BIO_push(b64.release(), bmem.release());
     BIO_set_flags(result, BIO_FLAGS_BASE64_NO_NL);
